fix: 달력 컴포넌트 네이밍 수정
